% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNumberLine.r
\name{ordinalNumberLine}
\alias{ordinalNumberLine}
\title{This function simulates the numberLine process and return the predited estimates.}
\usage{
ordinalNumberLine(
  targets,
  upperBound,
  lowerBound = 0,
  firstEstimate = NULL,
  memoryLength = NULL,
  rangeLength = 1,
  accuracyPercent = 0,
  numberSensitivity = 1,
  pIncludeConceptualPoints = 0,
  visibleReferencePoints = NULL,
  conceptualReferencePoints = NULL,
  targetOrder = "random",
  verbose = FALSE
)
}
\arguments{
\item{targets}{A vector of numbers that are the to-be-estimated values.}

\item{upperBound}{A number that identifes the upper point beyond which the partcipant cannot respond.  This is the upperBound of the bounded number line or the upper screen edge when the unbounded number line is used. This must be specified in "number-line units."}

\item{lowerBound}{A number that identifes the lower point beyond which the partcipant cannot respond.  This is the lowerBound of the bounded and unbounded number line or the lower screen edge when the universal number line is used.  DEFAULT = 0.}

\item{firstEstimate}{A proportion between 0 and 1 that idenitifies the bias of the first estimated value. The firstEstimate will always fall in the region between the visible reference points (or upper and lower bound) that retain ordinality. The first estmate is a value between 0-1, where 0 specifies the lowest point in that region, 1 specifies the highest point in that region, etc. This value will influence the remaining results.  If NULL, then a random draw from between relevant region will serve as the first estimate.  DEFAULT = NULL}

\item{memoryLength}{A integer that specifies the number of previous estimates that are remembered. The larger the number, the more accurate the estimates are going to be (given the constraints of the process). DEFAULT = legnth(targets).}

\item{rangeLength}{An integer that specifies the size of the window on the vector of remembered estimates that is used to identify the bounds of the next estimate. smaller number will produce more accurate estmates.  DEFAULT = 2}

\item{accuracyPercent}{A proportion between 0 and 1 that specifies the weight given to accurate responding: 0 = no weight, 1 = perfect responding.  DEFAULT = 0}

\item{numberSensitivity}{A proportion between 0 and 1 that specifies the range of equivelance. When numberSensitivity=0, the target is not differentiated from the nearest referenece point. When numberSensitivity=1, the target is differentiated from all other targets. When 0< numberSensitivity < 1, the range of equivelence is proportional to the value. DEFAULT = 1 (perfect precision)}

\item{visibleReferencePoints}{A vector of numbers that specify the visible (displayed) points that identify the value of positions on the number line. These are the upper and lower bound of the bounded number line, 0 and 1 for the unbounded number line, and anything the researcher uses for the universal number line or learning tasks.  Default is NULL.}

\item{conceptualReferencePoints}{A vector of numbers that specify any conceptual points that the participant may use to identify the value of positions on the number line. These may be the middle of the bounded number-line, or a learned position on the number_line. Currently, conceptualReferencePoints are only modeled if they are used thoughout the task (not introduced part of the way through) Default is NULL.}

\item{targetOrder}{A string specifying whether to keep the order in targets fixed ("fixed"), to ranndomize the order for every iteration of the loop ("random"), or to randomize it once and then use that order for all the loops ("single").  Default is "random".}

\item{verbose}{A boolean that specifies whether to print intermediate steps. This is used for debugging.  Default is FALSE.
'}
}
\value{
a dataframe containing the target value (target) and predicted estimate (fEst).
}
\description{
Function that simulates the numberLine process and return the predited estimates.
This function assumes that participants randomly place thier first response. Then their section response is randomly
presented between the first response and either the upper or lower boundary (depending on whether the target is greater
than or less than (respectively) the first target). Then, the partcipant will place the following estimates  ordinally
in relation to the previous estimates (and the upper and lower bounds). This is moderated by the memoryLength, which
specifies how well they remember the previous estimates (larger is better) and rangeLength, which specifies
how close they set the bounds around the remembered items (smaller is better). Finally, accuracyPercent specifies
the influence of accuracy on the the estimates.  This is a proportion, between 0 and 1, whereby estimate is
calculated as follows: estimate = (estimate based on algroithm * 1-accuracyPercent) + presented target * accuracyPercent
}
\examples{
ordinalNumberLine (c(2,3,4, 5, 6), 100, 0, firstEstimate = 400, rangeLength = 5, accuracyPercent = 0.5)
}
\keyword{line}
\keyword{number}
\keyword{ordinalNumberline}
