% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNumberLineFlex.r
\name{ordinalNumberLineFlex}
\alias{ordinalNumberLineFlex}
\title{This function runs the ordinalNumberLine with some flexible parameters.}
\usage{
ordinalNumberLineFlex(myTargets, parList, loops = 1000, verbose = FALSE)
}
\arguments{
\item{parList}{A list with the parameter names and values for the nlFit analysis. The list element name must be the parameter name and the value is the contents. The required elements are: firstEstimate; upperBound; lowerBound; rangeLength; memoryLength; numberSensitivity; accuracyPercent; referencePoinsts; targetOrder.}

\item{loops}{A number specifying the number of loops that will be run in the  simulation when it calculates the estimates. Higher numbers produce more precise estimates, but also increase the time needed to converge on a solution.  Default = 1000.}

\item{verbose}{A boolean that specifies whether to print intermediate steps. This is used for debugging.  Default is FALSE.
'}

\item{targets}{A vector of numbers that are the to-be-estimated values.}
}
\value{
A dataframe containing the target value (target) and predicted estimate (fEst).
}
\description{
Function that runs the ordinalNumberLine with some flexible parameters
}
\examples{
ordinalNumberLineFles (c(2,3,4, 5, 6), 100, 0, firstEstimate = 400, rangeLength = 5, accuracyPercent = 0.5, targetOrder = "random", loops = 100)
}
\keyword{number-line}
\keyword{ordinal}
\keyword{ordinalNumberline}
