% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimateNL.r
\name{getEstimateNL}
\alias{getEstimateNL}
\title{This function simulates the numberLine process and return the predited estimate for a single trial.}
\usage{
getEstimateNL(
  data,
  trialNumber,
  trialCol = "trial",
  targetCol = "target",
  valueCol = "targetValue",
  lowCol = "low",
  highCol = "high",
  estimateCol = "estimate",
  memoryLength = NULL,
  rangeLength = 1,
  numberSensitivity = 1,
  accuracyPercent = 0,
  firstEstimate = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A dataframe with the following columns: trialNumber, targetValue, estimate. It is created in ordinalNumberline() and passed to this function.}

\item{trialNumber}{A number that identifes the trial number of the estimate to be made.}

\item{trialCol}{A string specifying the column of data that holds the trialNumbers.  DEFAULT = "trial".}

\item{targetCol}{A string specifying the column of data that holds the targetValues.  DEFAULT = "target".}

\item{valueCol}{A string specifying the column of data that holds the target values.  DEFAULT = "targetValue".}

\item{lowCol}{A string specifying the column of data that holds the lower value of the range of equivelent numbers derived from numberSensitivity.  DEFAULT = "low".}

\item{highCol}{A string specifying the column of data that holds the higher value of the range of equivelent numbers derived from numberSensitivity.  DEFAULT = "high".}

\item{estimateCol}{A string specifying the column of data that holds the estimated values.  DEFAULT = "estimate".}

\item{memoryLength}{A integer that specifies the number of previous estimates that are remembered. The larger the number, the more accurate the estimates are going to be (given the constraints of the process). DEFAULT = legnth(targets).}

\item{rangeLength}{An integer that specifies the size of the window on the vector of remembered estimates that is used to identify the bounds of the next estimate. smaller number will produce more accurate estmates.  DEFAULT = 1}

\item{numberSensitivity}{A proportion between 0 and 1 that specifies the precision that previous targets are remembered. Specifically, it is the range of equivelance, where all targets between a low of (numberSensitivity * target) and a high of (2-numberSensitivity) are treated as equivelent and will be put in the same place on the number-line. DEFAULT = 1 (perfect precision)}

\item{accuracyPercent}{A proportion between 0 and 1 that specifies the weight given to accurate responding: 0 = no weight, 1 = perfect responding.  DEFAULT = 0}

\item{firstEstimate}{A number that is the first estimated value.  This value will influence the remaining results.  If NULL, then a random draw from between the upperBound and lowerBound will serve as the first estimate.  DEFAULT = NULL}

\item{verbose}{A boolean that specifies whether to print intermediate steps. This is used for debugging.  Default is FALSE.
'}
}
\value{
a dataframe containing the target value (target) and predicted estimate (fEst).
}
\description{
Function simulates the numberLine process and return the predited estimate for a single trial
}
\examples{
ordinalNumberLine (c(2,3,4, 5, 6), 100, 0, firstEstimate = 400, rangeLength = 5, accuracyPercent = 0.5)
}
\keyword{line}
\keyword{number}
\keyword{ordinalNumberline}
