% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalNumberLineFlex.r
\name{ordinalNumberLineFlex}
\alias{ordinalNumberLineFlex}
\title{This function runs the ordinalNumberLine with some flexible parameters.}
\usage{
ordinalNumberLineFlex(myTargets, parList, loops = 1000, verbose = FALSE)
}
\arguments{
\item{myTargets}{A vector of numbers that are the to-be-estimated values.}

\item{parList}{A list with the parameter names and values for the nlFit analysis. The list element name must be the parameter name and the value is the contents. The required elements are: firstEstimate; upperBound; lowerBound; memoryLength; numberSensitivity; accuracyPercent; pIncludeConceptualPoints; visibleReferencePoints; conceptualReferencePoints; targetOrder.}

\item{loops}{A number specifying the number of loops that will be run in the  simulation when it calculates the estimates. Higher numbers produce more precise estimates, but also increase the time needed to converge on a solution. Loops are collapsed to 1 when the run is deterministic, that is, when the target order is fixed or drawn once and there is no conceptual-point inclusion draw to average over.  Default = 1000.}

\item{verbose}{A boolean that specifies whether to print intermediate steps. This is used for debugging.  Default is FALSE.
'}
}
\value{
A dataframe containing the target value (target) and predicted estimate (fEst).
}
\description{
Function that runs the ordinalNumberLine with some flexible parameters
}
\examples{
ordinalNumberLineFlex(c(2, 3, 4, 5, 6),
                      parList = list(upperBound = 10, lowerBound = 0,
                                     firstEstimate = 0.5, memoryLength = 5,
                                     numberSensitivity = 1, accuracyPercent = 0.5,
                                     pIncludeConceptualPoints = 0,
                                     visibleReferencePoints = c(0, 10),
                                     conceptualReferencePoints = NULL,
                                     targetOrder = "fixed"),
                      loops = 10)
}
\keyword{number-line}
\keyword{ordinal}
\keyword{ordinalNumberline}
